create action creator fuction moving createanecdote and vote count to Reducer file
