+++ part6/redux-anecdotes/src/App.js
@@ -1,33 +1,18 @@
 import { useSelector, useDispatch } from 'react-redux'
 
+import { createNewAnecdote, addVote } from './reducers/anecdoteReducer'
+
 const App = () => {
   const anecdotes = useSelector((state) => state)
   const dispatch = useDispatch()
 
-  //  add votes
-  const addVote = (id) => {
-    return {
-      type: 'VOTE',
-      payload: { id },
-    }
-  }
   // add new anecdotes
 
   const addAnecdotes = (event) => {
     event.preventDefault()
     const content = event.target.anecdote.value
     event.target.anecdote.value = ''
-
-    const generateId = () => Number((Math.random() * 1000000).toFixed(0))
-
-    dispatch({
-      type: 'NEW_ANECDOTES',
-      payload: {
-        content,
-        id: generateId(),
-        votes: 0,
-      },
-    })
+    dispatch(createNewAnecdote(content))
   }
 
   let sortVotes = (a, b) => b.votes - a.votes
